# api/openapi/bundled.yaml
openapi: 3.0.0
info:
  title: Scene Hunter API
  description: Scene Hunter API
  version: v1
...
paths:
...
  /bucket/{prefix}:
    delete:
      operationId: Images_deleteBucket
      summary: バケット内のファイルを全て削除する
      description: バケット内のファイルを全て削除する
      parameters:
        - name: prefix
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.SuccessResponse'
        '400':
          description: 400 Bad Request エラーレスポンス
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.BadRequestResponse'
        '401':
          description: 401 Unauthorized エラーレスポンス
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.UnauthorizedResponse'
        '404':
          description: 404 Not Found エラーレスポンス
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.NotFoundResponse'
        '500':
          description: 500 Internal Server Error エラーレスポンス
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.InternalServerErrorResponse'
      tags:
        - Images
      security:
        - BearerAuth: []
...
  /file/{path}:
    get:
      operationId: Images_getFile
      summary: ファイルを取得する
      description: ファイルを取得する
      parameters:
        - name: path
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        '400':
          description: 400 Bad Request エラーレスポンス
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.BadRequestResponse'
        '401':
          description: 401 Unauthorized エラーレスポンス
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.UnauthorizedResponse'
        '404':
          description: 404 Not Found エラーレスポンス
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.NotFoundResponse'
        '500':
          description: 500 Internal Server Error エラーレスポンス
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.InternalServerErrorResponse'
      tags:
        - Images
      security:
        - BearerAuth: []
    delete:
      operationId: Images_deleteFile
      summary: ファイルを削除する
      description: ファイルを削除する
      parameters:
        - name: path
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.SuccessResponse'
        '400':
          description: 400 Bad Request エラーレスポンス
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.BadRequestResponse'
        '401':
          description: 401 Unauthorized エラーレスポンス
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.UnauthorizedResponse'
        '404':
          description: 404 Not Found エラーレスポンス
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.NotFoundResponse'
        '500':
          description: 500 Internal Server Error エラーレスポンス
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.InternalServerErrorResponse'
      tags:
        - Images
      security:
        - BearerAuth: []
...
  /list:
    get:
      operationId: Images_listFiles
      summary: ファイル一覧を取得する
      description: ファイル一覧を取得する
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.FileListResponse'
        '400':
          description: 400 Bad Request エラーレスポンス
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.BadRequestResponse'
        '401':
          description: 401 Unauthorized エラーレスポンス
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.UnauthorizedResponse'
        '500':
          description: 500 Internal Server Error エラーレスポンス
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.InternalServerErrorResponse'
      tags:
        - Images
      security:
        - BearerAuth: []
...
  /upload:
    post:
      operationId: Images_uploadImage
      summary: 画像をアップロードする
      description: 画像をアップロードする
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.ImageUploadResponse'
        '400':
          description: 400 Bad Request エラーレスポンス
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.BadRequestResponse'
        '401':
          description: 401 Unauthorized エラーレスポンス
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.UnauthorizedResponse'
        '500':
          description: 500 Internal Server Error エラーレスポンス
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.InternalServerErrorResponse'
      tags:
        - Images
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Routes.FormDataRequest'
      security:
        - BearerAuth: []
components:
  schemas:
...
    Models.BadRequestResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          enum:
            - invalid_request
            - invalid_parameter
            - missing_parameter
          description: エラーコード
        message:
          type: string
          description: エラーメッセージ
      description: 400 Bad Request エラーレスポンス
...
    Models.FileListResponse:
      type: object
      required:
        - objects
        - truncated
      properties:
        objects:
          type: array
          items:
            $ref: '#/components/schemas/Models.FileObject'
          description: オブジェクトリスト
        truncated:
          type: boolean
          description: 切り捨てられたかどうか
      description: ファイル一覧レスポンス
    Models.FileObject:
      type: object
      required:
        - key
        - uploaded
        - size
        - etag
      properties:
        key:
          type: string
          description: キー
        uploaded:
          type: string
          description: 最終更新日時
        size:
          type: integer
          format: int32
          description: サイズ
        etag:
          type: string
          description: etag
      description: ファイルオブジェクト
...
    Models.ImageUploadResponse:
      type: object
      required:
        - path
      properties:
        path:
          type: string
          description: ストレージパス
      description: 画像アップロードレスポンス
    Models.InternalServerErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          enum:
            - internal_error
          description: エラーコード
        message:
          type: string
          description: エラーメッセージ
      description: 500 Internal Server Error エラーレスポンス
...
    Models.NotFoundResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          enum:
            - not_found
          description: エラーコード
        message:
          type: string
          description: エラーメッセージ
      description: 404 Not Found エラーレスポンス
...
    Models.SuccessResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          description: 成功したかどうか
      description: 成功レスポンス
...
    Models.UnauthorizedResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          enum:
            - unauthorized
          description: エラーコード
        message:
          type: string
          description: エラーメッセージ
      description: 401 Unauthorized エラーレスポンス
...
    Routes.FormDataRequest:
      type: object
      required:
        - contentType
        - body
      properties:
        contentType:
          type: string
          enum:
            - multipart/form-data
        body:
          type: object
          properties:
            image:
              type: object
            w:
              type: integer
              format: int32
            h:
              type: integer
              format: int32
            q:
              type: integer
              format: int32
            f:
              type: string
            filename:
              type: string
          required:
            - image
            - filename
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: Bearer
servers:
  - url: https://scene-hunter.yashikota.com/api
    description: prd
    variables: {}
  - url: http://localhost:4282
    description: dev
    variables: {}